refactor: 콘솔 입력 메서드에 prompt 기능 추가
