+++ library-management/src/main/java/devcourse/backend/repository/FileRepository.java
@@ -1,4 +1,4 @@
-package devcourse.backend.medel;
+package devcourse.backend.repository;
 
 import devcourse.backend.medel.Book;
